requirements.txt
streamlit
pandas
openpyxl
numpy
psycopg2-binary
python-dotenv
sqlalchemy
